requirements_dev.txt
tensorflow
dvc
pandas
notebook
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure
joblib

-e .

# Dev requirements
pytest == 7.1.3
tox == 3.25.1
black == 22.8.0
flake8 == 5.0.4
mypy == 0.971